12868 Use presentationStyle enum to group events
