+++ WooCommerce/Classes/POS/Presentation/Card Present Payments/Connection Alerts/PointOfSaleCardPresentPaymentAlertType.swift
@@ -16,93 +16,3 @@ enum PointOfSaleCardPresentPaymentAlertType {
     case connectingFailedUpdateAddress(viewModel: PointOfSaleCardPresentPaymentConnectingFailedUpdateAddressAlertViewModel)
     case connectingFailedUpdatePostalCode(viewModel: PointOfSaleCardPresentPaymentConnectingFailedUpdatePostalCodeAlertViewModel)
 }
-
-extension CardPresentPaymentEventDetails {
-    func toAlertType() -> PointOfSaleCardPresentPaymentAlertType? {
-        switch self {
-        case .scanningForReaders(let endSearch):
-            return .scanningForReaders(
-                viewModel: PointOfSaleCardPresentPaymentScanningForReadersAlertViewModel(endSearchAction: endSearch))
-
-        case .scanningFailed(let error, let endSearch):
-            return .scanningFailed(
-                viewModel: PointOfSaleCardPresentPaymentScanningFailedAlertViewModel(
-                    error: error,
-                    endSearchAction: endSearch))
-
-        case .bluetoothRequired(let error, let endSearch):
-            return .bluetoothRequired(
-                viewModel: PointOfSaleCardPresentPaymentBluetoothRequiredAlertViewModel(
-                    error: error,
-                    endSearch: endSearch))
-
-        case .connectingToReader:
-            return .connectingToReader(viewModel: PointOfSaleCardPresentPaymentConnectingToReaderAlertViewModel())
-
-        case .connectingFailed(let error, let retrySearch, let endSearch):
-            return .connectingFailed(
-                viewModel: PointOfSaleCardPresentPaymentConnectingFailedAlertViewModel(
-                    error: error,
-                    retryButtonAction: retrySearch,
-                    cancelButtonAction: endSearch))
-
-        case .connectingFailedNonRetryable(let error, let endSearch):
-            return .connectingFailedNonRetryable(
-                viewModel: PointOfSaleCardPresentPaymentConnectingFailedNonRetryableAlertViewModel(
-                    error: error,
-                    cancelAction: endSearch))
-
-        case .connectingFailedUpdatePostalCode(let retrySearch, let endSearch):
-            return .connectingFailedUpdatePostalCode(
-                viewModel: PointOfSaleCardPresentPaymentConnectingFailedUpdatePostalCodeAlertViewModel(
-                    retryButtonAction: retrySearch,
-                    cancelButtonAction: endSearch))
-
-        case .connectingFailedChargeReader(let retrySearch, let endSearch):
-            return .connectingFailedChargeReader(
-                viewModel: PointOfSaleCardPresentPaymentConnectingFailedChargeReaderAlertViewModel(
-                    retryButtonAction: retrySearch,
-                    cancelButtonAction: endSearch))
-
-        case .connectingFailedUpdateAddress(let wcSettingsAdminURL, let retrySearch, let endSearch):
-            return .connectingFailedUpdateAddress(
-                viewModel: PointOfSaleCardPresentPaymentConnectingFailedUpdateAddressAlertViewModel(
-                    settingsAdminUrl: wcSettingsAdminURL,
-                    retrySearchAction: retrySearch,
-                    cancelSearchAction: endSearch))
-
-        case .foundReader(let name, let connect, let continueSearch, let endSearch):
-            return .foundReader(
-                viewModel: PointOfSaleCardPresentPaymentFoundReaderAlertViewModel(
-                    readerName: name,
-                    connectAction: connect,
-                    continueSearchAction: continueSearch,
-                    endSearchAction: endSearch))
-
-        case .updateProgress(let requiredUpdate, let progress, let cancelUpdate):
-            return .updatingReader(viewModel: CardPresentPaymentUpdatingReaderAlertViewModel())
-
-        case .updateFailed(let tryAgain, let cancelUpdate):
-            return .updateFailed(viewModel: CardPresentPaymentReaderUpdateFailedAlertViewModel())
-
-        case .updateFailedNonRetryable(let cancelUpdate):
-            return .updateFailed(viewModel: CardPresentPaymentReaderUpdateFailedAlertViewModel())
-
-        case .updateFailedLowBattery(let batteryLevel, let cancelUpdate):
-            return .updateFailed(viewModel: CardPresentPaymentReaderUpdateFailedAlertViewModel())
-
-            /// Payment messages are not handled as alerts, so should not be converted to alert view models
-        case .preparingForPayment,
-                .selectSearchType,
-                .tapSwipeOrInsertCard,
-                .success,
-                .error,
-                .errorNonRetryable,
-                .processing,
-                .displayReaderMessage,
-                .cancelledOnReader,
-                .validatingOrder:
-            return nil
-        }
-    }
-}
